add i64 encoding/decoding for dub sid-eid-span scheme

+++ midi_timidity.c
@@ -83,7 +83,7 @@ TIMIDITY_Shutdown(
 static BOOL
 TIMIDITY_Play(
 	VOID       *object,
-	INT         iNum,
+	int64_t     iNum,
 	BOOL        fLoop,
 	FLOAT       flFadeTime
 	)
